Rename components to camelCase
